+++ src/plugins/security.ts
@@ -5,7 +5,7 @@ import _securityPlugin from 'eslint-plugin-security'
 // https://github.com/eslint-community/eslint-plugin-security#readme
 export const securityPlugin = definePlugin({
   name: 'security',
-  settings: {},
+  source: _securityPlugin,
   // https://github.com/eslint-community/eslint-plugin-security/tree/main/rules
   rules: extendPluginRules('security', _securityPlugin.configs.recommended),
   testRules: extendPluginRules(